feat: New arch API and commonjs(#13)
* doc(shared): improve ZenroomError doc

* doc(shared): rm obsolete line

I wanted to freeze the object, but decided that it wasn't worth it, but
apparently forgot to remove the comment.

* refactor(ignored): explicitly mark the return type

I decided that it is more clear this way.

* doc(ignored): clarify usage

Clarify the usage of getIgnoredStatements().

* refactor(shared)!: rename ZenroomResult type

Rename it to ZenroomOutput.

* feat(shared)!: parse zencode_exec's output result

Convert the JSON string to a JS object before we return it.

* Test(shared): fix the description

Fix the name of the function being tested in the description.

* test(shared): improve zenroom test

I think this better captures my intention.

* refactor(fs): shut linter down

Add missing no-non-null-assertion comments.

* style(fs): spread array into lines

* style(fs): merge lines of an array

* refactor(ignored): return zenroom output as well

Return the Zenroom output along with the ignored statements in
getIgnoredStatements.

* test(ignored): apply getIgnoredStatements() changes

Due to the recently-introduced changes to getIgnoredStatements(), we apply
changes to tests as well.

* test(fs): apply getIgnoredStatements() changes

Due to the recently-introduced changes to getIgnoredStatements(), we apply
changes to tests as well.

* feat(shared): add new tokens

Add the tokens:

- I
- Given
- When
- Then

each of which, except for I, is case-insensitive, as also done in
Zenroom.

* refactor(ignored): explicitly mark with type

Explicitly mark a TypeScript type with the "type" keyword.

* refactor(fs)!: switch to shared tokens

We now use the shared Then and I.  I also removed And statements due to
their complexity.

* test(fs): rm And statements

They are too complex to deal with right just now.

* refactor(ignored)!: revert back changes

Revert back changes to getIgnoredStatements().  I had planned the API little bit
differently at first.  I thought I should get the execution result along with
the ignored statements, but it wasn't quite right.  What needs to be done
follows like this:

- Get ignored statements
- Decide whether the individual statements must run before or after the actual
  execution
- Execute the before statements and save the output
- Execute the "actual execution" using the provided data and the data you get
  from executing the before statements using the provided data and the data you
  get from executing the before statements merged together, and save the output
- Execute the after statements, merge the output with the data you get from the
  actual execution, and save it
- Return the output to the API consumer.

* fix(ignored): apply changes to ignored api

Due to I reverted my previous changes, we need to apply previous changes back!

* fix(fs): apply changes to ignored api

Due to I reverted my previous changes, we need to apply previous changes back!

* test(fs): add keyholder test case

This test case supposed to test for creating public/secret keypairs and saving
them to files.  It was a bit complicated to do so, so we are generating random
string and saving that to a file instead.

* chore(git): rm unnecessary line in gitignore

The previous *.ts includes *.test.ts files, so the latter was
unnecessary.

* refactor(*)!: better structure pkgs

* feat(*): support commonjs

* feat(*): build with tslib included

Because we don't wanna require each and every library user of us to add
tslib to their dependencies.

* fix(shared): rm unused export

* fix(*): fix cjs not working

This requires each subpackage to have another package.json changing its type.

* feat: add core module

This module introduces a method to use plugins defined in subpackages, and use
that to create your own dialect of Zencode.

* feat: convert fs to use core

Convert the fs module and its example(s) to use the recently-introduced core
module.

* style: don't specify uncessary type