feat: Enforce order of the statements
Slangroom statement will be exectuted in the order they are written
such that it is possible to use data from a previous computation.

All slangroom statement should be written before (for given) or
after (for then) the zenroom statements. This is because, at the
moment, we cannot start executing zenroom and suspend its execution
before each slangroom statement (maybe it will be possible with
improvement on exportable heap in zenroom).