feat!: Implement new core: new lexer/parser/visitor/plugin systems (#40)
* feat(fs): introduce the fs plugin

* fix(fs): typo

* feat(fs): convert sandboxdir into a function

* fix(fs): fix return values

* feat(fs): finish up and add tests

* feat(git): introduce the git module

* feat!(core,http,ethereum): unify the parser

* feat(core)!: introduce new lexer/parser/visitor

This new core here eases the heavy work on the developer by simplifing the
API while keeping all the things we want.  At first, we used to use Chevrotain
with a method that didn't allow method that didn't allow us to detect the
individual syntax errors of a custom statement we defined.  I worked around that
issue by casting black-magic spells over Chevrotain internals.  I hated that
solution, but at that time, it was the most practicle one.

Now, here, I introduce the new core with its new lexer/parser/visitor that is
smart enough to cover most of our real-world needs.  The visitor is in perfect
Shape shape.  The lexer is almost perfect, perhaps needs just a little bit of
polishing, but it is the one that is the simplest.  The parser is smart enough
to cover most of our real-world cases, but it also needs some treatment.  I'm
planning to have a ranking system over it.  Currently, it uses the number of
errors detected on a statement, and using that, it ranks its matches.  This
solution appears to be quite reliable in practice.

The way we use the API is quite simple ("Given I" parts can also be "Then I"):

	import {Plugin} from '@slangroom/core'

	const p = new Plugin();
	const cb = (ctx) => ctx.fail('example') // cb is called when statement is matched
	p.new('love asche', cb)
	// -> Given I love Asche

	p.new('open', 'read file contents', cb)
	// -> Given I open 'ident' and read file contents

	p.new('connect', 'send http request', cb)
	// -> Given I connect to 'ident' and send http request

	p.new(['base32'], 'convert to base64', cb)
	// -> Given I send base32 'ident' and convert to base64

	p.new('connect', ['object', 'proxy'], 'send http request', cb)
	// -> Given I connect to 'ident' and send object 'ident' and send proxy 'ident' and send http request

	export const myPlugin = p;

* feat(ethereum)!: port over to the new core and fix tests

* feat(http)!: port over to the new core and fix tests

* fix(core): relax phrase and params checks

* feat(wallet)!: port over to the new core

Currently, the tests fail for some reason.

* feat(core)!: forbid spaces in params

* fix(wallet)!: switch over to underscore params

* fix: wrongo order of the bindings set /wor/ident/

* fix: the wallet params names

* lint: linting

* fix(core): fix parsing of params order

* test(ethereum): fix rule ignore

* fix(wallet): parameter names

---------

Co-authored-by: Puria Nafisi Azizi <puria@dyne.org>