fix!(http): change the output format
Status codes used to be output using integers, but due to some bug fix in
Zenroom, a 'string dictionary' now converts numbers to strings.  For that very
reason, we decided that the best decision to take here is to use strings instead
of integers everywhere.

Along with this, I packed in some simplifications and best-to-dos by axios docs.