feat(grammar): update autocompletion with new format (#226)
* feat(grammar): update autocompletion with new format

* feat: first draft for autogenerated grammar syntax

improved strictness on grammar, but at the same time we lost highlight of the core part of the statement. Need some time more to understand how to get both. Moreover this cover only old format, once it is ready I should include also new format

* fix: highlight of statement core part

* fix: add syntax.grammar.template file

* fix: add new format grammar

* chore(grammar): add parametrized

soon or later also this part will be filled programmatically

* test(grammar): update test output

* feat(grammar): autogenerate also tokens

* fix: add output secret variant

* fix: add missing secret word

* WIP

* refactor(wallet): substitute p-256 with p256

minus sign are not allowed

* fix(grammar): autogenerate also the list of tokens

* fix(grammar): add missing tokens template

* chore: add prepare, compute, with, where and secret words to highlight